requirements.txt
respeecher
deepl
pysubs2
whisperx
piper-tts
